monitoring-app/requirements.txt
mysql-connector-python
fastapi
uvicorn
psycopg2-binary